+++ weaver/wps_restapi/quotation/__init__.py
@@ -30,11 +30,11 @@ def includeme(config):
     config.add_route(**sd.service_api_route_info(sd.quote_service, settings))
     config.add_route(**sd.service_api_route_info(sd.bills_service, settings))
     config.add_route(**sd.service_api_route_info(sd.bill_service, settings))
-    config.add_view(q.get_quote_estimator, route_name=sd.process_estimator_service.name,
+    config.add_view(q.get_process_quote_estimator, route_name=sd.process_estimator_service.name,
                     request_method="GET", renderer=OutputFormat.JSON)
-    config.add_view(q.update_quote_estimator, route_name=sd.process_estimator_service.name,
+    config.add_view(q.update_process_quote_estimator, route_name=sd.process_estimator_service.name,
                     request_method="PUT", renderer=OutputFormat.JSON)
-    config.add_view(q.delete_quote_estimator, route_name=sd.process_estimator_service.name,
+    config.add_view(q.delete_process_quote_estimator, route_name=sd.process_estimator_service.name,
                     request_method="DELETE", renderer=OutputFormat.JSON)
     config.add_view(q.get_quote_list, route_name=sd.quotes_service.name,
                     request_method="GET", renderer=OutputFormat.JSON)